# python/backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Weekly Progress API",
    "version": "0.1.0"
  },
  "paths": {
    "/api/health": {
      "get": {
        "tags": [
          "weekly"
        ],
        "summary": "Health",
        "operationId": "health_api_health_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Health Api Health Get"
                }
              }
            }
          }
        }
      }
    },
    "/api/weekly-drafts/generate": {
      "post": {
        "tags": [
          "weekly"
        ],
        "summary": "Generate Weekly Drafts",
        "description": "Generate two drafts:\n- Weekly progress (copy/paste to Confluence)\n- Weekly manager review (copy/paste to Confluence)",
        "operationId": "generate_weekly_drafts_api_weekly_drafts_generate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GenerateWeeklyDraftsRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenerateWeeklyDraftsResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/weekly-drafts": {
      "get": {
        "tags": [
          "weekly"
        ],
        "summary": "List Weekly Drafts",
        "operationId": "list_weekly_drafts_api_weekly_drafts_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/WeeklyDraftListItem"
                  },
                  "type": "array",
                  "title": "Response List Weekly Drafts Api Weekly Drafts Get"
                }
              }
            }
          }
        }
      }
    },
    "/api/weekly-drafts/{draft_id}": {
      "get": {
        "tags": [
          "weekly"
        ],
        "summary": "Get Weekly Draft",
        "operationId": "get_weekly_draft_api_weekly_drafts__draft_id__get",
        "parameters": [
          {
            "name": "draft_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Draft Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WeeklyDraftDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "weekly"
        ],
        "summary": "Update Weekly Draft",
        "operationId": "update_weekly_draft_api_weekly_drafts__draft_id__put",
        "parameters": [
          {
            "name": "draft_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Draft Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateWeeklyDraftRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WeeklyDraftDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "GenerateWeeklyDraftsRequest": {
        "properties": {
          "week_start": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Week Start"
          },
          "week_end": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Week End"
          },
          "jira_days_lookback": {
            "type": "integer",
            "maximum": 90.0,
            "minimum": 1.0,
            "title": "Jira Days Lookback",
            "default": 14
          },
          "confluence_page_ids": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Confluence Page Ids"
          },
          "save_to_db": {
            "type": "boolean",
            "title": "Save To Db",
            "default": true
          },
          "save_to_files": {
            "type": "boolean",
            "title": "Save To Files",
            "default": true
          }
        },
        "type": "object",
        "title": "GenerateWeeklyDraftsRequest",
        "description": "Generate drafts for a given week window.\n\n- If `week_start/week_end` are omitted, backend computes the current week (Mon-Sun) window.\n- `jira_days_lookback` controls which tickets to fetch (default: 14, matches existing MCP tools)."
      },
      "GenerateWeeklyDraftsResponse": {
        "properties": {
          "week_start": {
            "type": "string",
            "format": "date",
            "title": "Week Start"
          },
          "week_end": {
            "type": "string",
            "format": "date",
            "title": "Week End"
          },
          "jira_days_lookback": {
            "type": "integer",
            "title": "Jira Days Lookback"
          },
          "drafts": {
            "items": {
              "$ref": "#/components/schemas/GeneratedDraft"
            },
            "type": "array",
            "title": "Drafts"
          }
        },
        "type": "object",
        "required": [
          "week_start",
          "week_end",
          "jira_days_lookback",
          "drafts"
        ],
        "title": "GenerateWeeklyDraftsResponse"
      },
      "GeneratedDraft": {
        "properties": {
          "kind": {
            "type": "string",
            "title": "Kind"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "content_markdown": {
            "type": "string",
            "title": "Content Markdown"
          },
          "saved_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Saved Id"
          },
          "saved_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Saved Path"
          }
        },
        "type": "object",
        "required": [
          "kind",
          "title",
          "content_markdown"
        ],
        "title": "GeneratedDraft"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "UpdateWeeklyDraftRequest": {
        "properties": {
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "content_markdown": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Content Markdown"
          }
        },
        "type": "object",
        "title": "UpdateWeeklyDraftRequest"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "WeeklyDraftDetail": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "kind": {
            "type": "string",
            "title": "Kind"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "week_start": {
            "type": "string",
            "format": "date",
            "title": "Week Start"
          },
          "week_end": {
            "type": "string",
            "format": "date",
            "title": "Week End"
          },
          "content_markdown": {
            "type": "string",
            "title": "Content Markdown"
          }
        },
        "type": "object",
        "required": [
          "id",
          "kind",
          "title",
          "week_start",
          "week_end",
          "content_markdown"
        ],
        "title": "WeeklyDraftDetail"
      },
      "WeeklyDraftListItem": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "kind": {
            "type": "string",
            "title": "Kind"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "week_start": {
            "type": "string",
            "format": "date",
            "title": "Week Start"
          },
          "week_end": {
            "type": "string",
            "format": "date",
            "title": "Week End"
          }
        },
        "type": "object",
        "required": [
          "id",
          "kind",
          "title",
          "week_start",
          "week_end"
        ],
        "title": "WeeklyDraftListItem"
      }
    }
  }
}